Jacobian 1.0
I tried to use the implementation given on the slides from lecture 19. I'm not sure if this works, as it seems very different from the other numerical version we did. Although it seems like this could work as the first slide in lecture 19 mentions the numerical way of calculating the Jacobian as distinct from this method.